Launch GitUp via the command line tool
Opening GitUp with the `open` command works well for projects where the
repository root and the VSCode workspace root path are the same. It
doesn’t work for projects where the workspace root is a child of the
repository root. The command line tool, included with GitUp, is built to
handle this sort of situation gracefully.